Improve memory usage of tracking line numbers of bytes
Since multiple instructions can come from the same line of source code,
we can track just the byte offset and line number of the first byte of that
line.

Then when we want to print out the line number, we look for the closest
offset via binary search and return the line number that the closest offset
appears on.

In the initial implementation, we tracked the line number for every byte.

This new way reduces the amount of data we have to record thereby improving
memory usage.